feat: pulling inspector into a seperator package, and revamping how types are used
This is a major refactor (that's a work in progress). The goal is to allow us to be able to differentiate between different import paths with the same name.

The main issue is the typescript schema generator still doesnt allow duplicates, so it's not considered complete.

The second feature, which is much bigger, is pulling the inspector out into a different library. This means we can simplify the serialization logic quite a bit and have our unit tests (or end to end tests) run by having example projects and chekcing out the generated configuration used to run vramework.

This will likely be turned a bit inside out as well in order to optimise compiling in the future. If we store everything based by file name we could hash them which would allow us to skip files we already know and haven't changed. The main issue would be the cascading effects, so we would need to propogate changes.

Either ways, it's an optimization effort, and currently is not something we need!